+++ src/scripts/domains/extensionButton.ts
@@ -1,7 +1,7 @@
 import { waitElement } from '@1natsu/wait-element'
 import { NodeLike } from '@1natsu/wait-element/dist/types/options'
 import { extensionSwitcher } from '../controllers/extensionSwicher'
-import toggleTooltip from '../libs/toggleTooltip'
+import toggleTooltip from '../usecases/toggleTooltip'
 import { hasInjected, isActive } from '../repository/extensionState'
 
 /**
