+++ src/scripts/usecases/toggleTooltip.ts
